requirements.txt
streamlit==1.23.0
folium==0.12.1
pandas==1.5.3
requests==2.31.0
geopy==2.3.0
python-dotenv==1.0.0